refactor: move `MockStateManager` to `mpc-chain-integration-core` and use in both Ethereum and Solana indexer tests
