+++ chain-signatures/chain-solana/Cargo.toml
@@ -29,3 +29,4 @@ tracing.workspace = true
 
 [dev-dependencies]
 mockito.workspace = true
+tracing-subscriber.workspace = true
